+++ lib/browser/puppeteer_wrapper/puppeteer_utils.ts
@@ -1,4 +1,4 @@
-import { ConsoleMessage, JSHandle } from '../browser/puppeteer_wrapper/puppeteer_types';
+import { ConsoleMessage, JSHandle } from './puppeteer_types';
 
 export async function stringifyLogText(log: ConsoleMessage): Promise<string> {
     const text = log.text();
